ADDED: Remaining planets

+++ Planets/Models/AbstractClasses/InterstarPlanet.cs
@@ -3,6 +3,6 @@
     // Note - Interstar planet are part of a solar system - as opposite to exostar planets
     public abstract class InterstarPlanet : Planet
     {
-        public Tuple<int, int> DistanceFromSun { get; set; } = Tuple.Create(0, 0);
+        public Tuple<long, long> DistanceFromSun { get; set; } = Tuple.Create((long)0, (long)0);
     }
 }
